non working cpu logic
main.tscn is giving me a problem with merges from switching branches. added CPUDeck and CPUHand along with the other components of each.  calling draw_cpu_card causes a problem placement of the card. with that line (14 in main_game) commented out, the game will run though there is not a call for the cpu_turn added in yet to start the cpu turn